make sure Rancher is ready
Gate the run of the rancher test cases behind a pseudo-test case that
just waits and passes once it detects that Rancher deployments are
available and able to service requests.
This is particularly important for the rancher-webhook, as Rancher may
start servicing requests without the webhook being ready. But when that
happens buggy behavior is introduced.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>